+++ CHANGELOG.md
@@ -1,44 +1,59 @@
 # Quarters Unity SDK changelog
 
 ## 2.0.0 (2022-05-30)
+
 - Rewrite of the SDK 1.0.0
 
-
 ## 2.0.1 (2022-05-30)
+
 - Bug fixes
 
 ## 2.0.2 (2022-06-06)
+
 - Added Windows Unity Editor authorization flow
 
 ## 2.1.0 (2022-06-07)
+
 - Code formatting fixes
 - Added Console logging level selection
 
 ## 2.1.1 (2022-06-07)
+
 - Added VSP attribution
 
 ## 2.1.2 (2022-06-20)
+
 - iOS post build process fix to missing domain association config
 - Unity new Input system is now supported
 
 ## 2.2.1 (2022-07-12)
+
 - Added [new QuartersSDK project (this project has all the logic](https://github.com/weiks/quarters-unity-sdk/tree/master/QuartersSDK)) reference
 - Logging using dependency injection
 - Parameters vars
 - Unit testing for QuartersSDK (inside new project)
 
 ## 2.2.2 (2022-08-31)
+
 - Added new test cases for QuartersSDK
 - Uploaded using last version of Unity Uploader
-  
+
 ## 2.2.3 (2022-09-21)
-- Added APP name while calling VSP 
+
+- Added APP name while calling VSP
 
 ## 2.3.2021 (2022-10-24)
+
 - Support for `netstandard 2.0`
 - Updated VSP log info
 - Version compatible with Unity 2021 and above ( if you are using old versions of Unity use the [release 2.3.2019](https://github.com/weiks/poq-unity-package-manager/releases/tag/2.3.2019) )
 
 ## 3.0.2021 (2022-10-24)
+
 - Replaced parameter `APP_KEY` with `code`
 - Version compatible with Unity 2021 and above ( if you are using old versions of Unity use the [release 3.0.2019](https://github.com/weiks/poq-unity-package-manager/releases/tag/3.0.2019) )
+
+## 3.1.2021 (2022-04-21)
+
+- Updated links from `QuartersInit` (`App Dashboard` and `My apps`)
+- Added Safari WebView support for iOS.